added support for templated pages
The templates are located in resources/pages, while snippets go to
resources/snippets. If a request contains a path that matches a page, the page
template will be loaded, processed and wrapped into a response. The response is
rendered in a last step (see wrap-response-handler) to make sure the html
content might be modified in other stages of the middleware chain.