added support for ClojureScript
There now is a middleware that will compile all ClojureScript code located in
the according directory. The compiler is called either on startup of the server
(in production mode) or whenever the code changed (in development mode).
Additionally there was some restructuring to the environment labels. They are
now :prod and :dev!